+++ nexus-server/src/handlers/chat.rs
@@ -147,7 +147,7 @@ where
 
         let timestamp = SystemTime::now()
             .duration_since(UNIX_EPOCH)
-            .map(|d| d.as_secs())
+            .map(|d| d.as_secs() as i64)
             .unwrap_or(0);
         let chat_message = ServerMessage::ChatMessage {
             session_id: id,
